+++ include/ui/tileseteditor.h
@@ -84,6 +84,8 @@ private slots:
     void on_actionShow_Unused_toggled(bool checked);
     void on_actionShow_Counts_toggled(bool checked);
     void on_actionShow_UnusedTiles_toggled(bool checked);
+    void on_actionMetatile_Grid_triggered(bool checked);
+    void on_actionLayer_Grid_triggered(bool checked);
 
     void on_actionUndo_triggered();