+++ include/ui/tileseteditormetatileselector.h
@@ -22,6 +22,7 @@ public:
     QVector<uint16_t> usedMetatiles;
     bool selectorShowUnused = false;
     bool selectorShowCounts = false;
+    bool showGrid;
 
 protected:
     void mousePressEvent(QGraphicsSceneMouseEvent*);
@@ -35,10 +36,12 @@ private:
     Tileset *secondaryTileset = nullptr;
     uint16_t selectedMetatile;
     int numMetatilesWide;
+    int numMetatilesHigh;
     uint16_t getMetatileId(int x, int y);
     QPoint getMetatileIdCoords(uint16_t);
     bool shouldAcceptEvent(QGraphicsSceneMouseEvent*);
-
+    int numRows(int numMetatiles);
+    int numRows();
     void drawFilters();
     void drawUnused();
     void drawCounts();