+++ src/ui/metatilelayersitem.cpp
@@ -4,7 +4,7 @@
 #include <QPainter>
 
 void MetatileLayersItem::draw() {
-    const QList<QPoint> tileCoords = QList<QPoint>{
+    static const QList<QPoint> tileCoords = QList<QPoint>{
         QPoint(0, 0),
         QPoint(16, 0),
         QPoint(0, 16),
@@ -19,8 +19,11 @@ void MetatileLayersItem::draw() {
         QPoint(80, 16),
     };
 
-    QPixmap pixmap(projectConfig.getNumLayersInMetatile() * 32, 32);
+    const int numLayers = projectConfig.getNumLayersInMetatile();
+    QPixmap pixmap(numLayers * 32, 32);
     QPainter painter(&pixmap);
+
+    // Draw tile images
     int numTiles = projectConfig.getNumTilesInMetatile();
     for (int i = 0; i < numTiles; i++) {
         Tile tile = this->metatile->tiles.at(i);
@@ -28,6 +31,14 @@ void MetatileLayersItem::draw() {
                 .mirrored(tile.xflip, tile.yflip)
                 .scaled(16, 16);
         painter.drawImage(tileCoords.at(i), tileImage);
+    }
+    if (this->showGrid) {
+        // Draw grid
+        painter.setPen(Qt::white);
+        for (int i = 1; i < numLayers; i++) {
+            int x = i * 32;
+            painter.drawLine(x, 0, x, 32);
+        }
     }
 
     this->setPixmap(pixmap);